Display glossary entry

Display glossary entry

Usage

glossary(
  term,
  display = term,
  def = NULL,
  add_to_table = TRUE,
  show = c("term", "def"),
  popup = glossary_popup(),
  path = glossary_path()
)

Arguments

term

The glossary term to link to, can contain spaces

display

The display (if different than the term)

def

The short definition to display on hover and in the glossary table; if NULL, this will be looked up from the glossary.yml file

add_to_table

whether to add to the table created by glossary_table()

show

whether to show the term or just the definition

popup

whether to show the popup on "click" or "hover" (or "none")

path

the path to the glossary file

Value

string with the link and hover text

Examples

glossary("alpha")
#> Warning: The definition for "alpha" was not found in 
#> [1] "<a class='glossary'>alpha<span class='def'></span></a>"
glossary("alpha", "$\\alpha$")
#> Warning: The definition for "alpha" was not found in 
#> [1] "<a class='glossary'>$\\alpha$<span class='def'></span></a>"
glossary("alpha", def = "The first letter of the Greek alphabet")
#> [1] "<a class='glossary'>alpha<span class='def'>The first letter of the Greek alphabet</span></a>"
glossary("alpha", show = "term")
#> Warning: The definition for "alpha" was not found in 
#> [1] "<a class='glossary'>alpha<span class='def'></span></a>"
glossary("alpha", show = "def")
#> Warning: The definition for "alpha" was not found in 
#> [1] ""