Load all definitions

Load all the definitions in a glossary file, usually for subsequent display using glossary_table

Usage

glossary_load_all(path = glossary_path())

Arguments

path

The path to the glossary file; set default with glossary_path

Value

NULL; Called for side effects

Examples

demo_glossary <- system.file("glossary.yml", package = "glossary")
glossary_load_all(demo_glossary)

glossary_table()
#> <table class="table" style="margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> term </th>
#>    <th style="text-align:left;"> definition </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> SESOI </td>
#>    <td style="text-align:left;"> Smallest Effect Size of Interest: the smallest effect that is theoretically or practically meaningful
#> 
#> See [Equivalence Testing for Psychological Research](https://doi.org/10.1177/2515245918770963) for a tutorial on methods for choosing an SESOI. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> alpha </td>
#>    <td style="text-align:left;"> The threshold chosen in Neyman-Pearson hypothesis testing to distinguish test results that lead to the decision to reject the null hypothesis, or not, based on the desired upper bound of the Type 1 error rate. An alpha level of 5% is most commonly used, but other alpha levels can be used as long as they are determined and preregistered by the researcher before the data is analyzed. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> alpha (graphics) </td>
#>    <td style="text-align:left;"> A value between 0 and 1 used to control the levels of transparency in a plot </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> effect size </td>
#>    <td style="text-align:left;"> 'quantitative reflection of the magnitude of some phenomenon that is used for the purpose of addressing a question of interest' (Kelley &amp; Preacher, 2012) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> html </td>
#>    <td style="text-align:left;"> This is a paragraph with a [link](https://url.com).
#> 
#> And another paragraph before a list:
#> 
#> * Item 1
#> * List 2 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> joins </td>
#>    <td style="text-align:left;"> Ways to combine data from two tables </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> p-value </td>
#>    <td style="text-align:left;"> The probability of the observed data, or more extreme data, if the null hypothesis is true. The lower the p-value, the higher the test statistic, and less likely it is to observe the data if the null hypothesis is true. </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> power </td>
#>    <td style="text-align:left;"> The probability of rejecting the null hypothesis when it is false, for a specific analysis, effect size, sample size, and criteria for significance. </td>
#>   </tr>
#> </tbody>
#> </table>